split put
